music /requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
midi2audio==0.1.1
mido==1.3.2
mingus @ git+https://github.com/victor823543/python-mingus.git@c154c07e5d527e21030484e19c4522eca602ff26
numpy==1.26.4
packaging==23.2
pretty_midi==0.2.10
psycopg2-binary==2.9.9
pydub==0.25.1
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
